Save user session to database

+++ server/user-service/domain/entity/session.go
@@ -5,7 +5,7 @@ type Sessions struct {
 	AccessToken  string
 	Email        string
 	RoleID       int
-	Exp          int
-	CreatedAt    int
-	UpdatedAt    int
+	Exp          int64
+	CreatedAt    int64
+	UpdatedAt    int64
 }
